refactor: import Singleton from utils.singleton directly
